[Analysis] Update changepoint input buffer from verdicts to runs.
Rather than storing a list of verdicts, update the input buffer
to store test runs only. Duplicate runs are not stored. This sets
us up for transitioning to a lower-latency data ingestion, which
ingests test runs as they complete rather than waiting for entire
verdicts to be complete.

The underlying changepoint analysis always used counts of runs as the
statistics for determining whether a changepoint existed; in this
respect no change is required. However, the existing implementation
reported some verdict level statistics which can no longer be
supported in their current form.

In the new implementation we instead report statistics for
'source verdicts', the aggregation of all test results at a source
position. This means that the flaky verdict count also captures
any cross-build flakiness at the same source position (including
cross-builder flakiness), not only in-build flakiness.
For most purposes clients will not notice a huge difference either
way and the documentation has been updated to reflect this change.
The field names for external protos have not been updated at this
stage to protect compatibility with existing clients.

To support the transition from storing verdicts to runs in the input
buffer, we keep a deserialiser for the legacy verdicts format, to
allow seamless transition with existing changepoint history in tact.

BUG=b:333652213
TEST=Integration tests

Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5568014
Commit-Queue: Patrick Meiring <meiring@google.com>

+++ analysis/internal/changepoints/sources/sources.go
@@ -52,6 +52,6 @@ func HasCommitData(sources *pb.Sources) bool {
 	return commit.GetHost() != "" && commit.GetProject() != "" && commit.GetRef() != "" && commit.GetPosition() != 0
 }
 
-func CommitPosition(sources *pb.Sources) int {
-	return int(sources.GitilesCommit.Position)
+func CommitPosition(sources *pb.Sources) int64 {
+	return sources.GitilesCommit.Position
 }
